webserver/requirements.txt
better-profanity==0.6.1
blinker==1.4
boto3==1.16.0
celery==5.0.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.14.0
Flask-SendGrid==0.7.1
Flask-SQLAlchemy==2.4.4
google-auth-oauthlib==0.4.1
marshmallow-sqlalchemy==0.23.1
numpy==1.19.2
pandas==1.1.3
paramiko==2.7.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-jose==3.2.0
pyzipcode==2.0.0
redis==3.5.3
requests==2.24.0
sentry-sdk==0.19.1
stripe==2.55.0
waitress==1.4.4
datadog==0.39.0